# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IIRP V1",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "summary": "Analysis",
        "operationId": "analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "result_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Result Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyses/{analysis_id}/export": {
      "get": {
        "summary": "Export",
        "operationId": "export_api_v1_analyses__analysis_id__export_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "result_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Result Ids"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(csv|json)$",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/limits": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Import Limits",
        "operationId": "import_limits_api_v1_events_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportLimitsOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/prompt": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Prompt",
        "operationId": "prompt_api_v1_events_prompt_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "custom",
                "earnings"
              ],
              "type": "string",
              "default": "custom",
              "title": "Kind"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "zh",
                "en"
              ],
              "type": "string",
              "default": "zh",
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPromptOutput"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/events/preview": {
      "post": {
        "tags": [
          "event research"
        ],
        "summary": "Preview",
        "operationId": "preview_api_v1_events_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPreviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPreviewOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/confirm": {
      "post": {
        "tags": [
          "event research"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_events_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConfirmInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConfirmOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/sets": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Sets",
        "operationId": "sets_api_v1_events_sets_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "custom",
                    "earnings"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSetsOutput"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/events/sets/{set_id}": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_events_sets__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSetOutput"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/events/sets/{set_id}/analyses": {
      "post": {
        "tags": [
          "event research"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis_api_v1_events_sets__set_id__analyses_post",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAnalysisInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAnalysisCreated"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/events/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Analysis",
        "operationId": "analysis_api_v1_events_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "result_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAnalysisOutput"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/events/analyses/{analysis_id}/export": {
      "get": {
        "tags": [
          "event research"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_events_analyses__analysis_id__export_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "result_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventErrorOutput"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOutput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AnalysisItem"
            },
            "type": "array",
            "title": "Results"
          },
          "batch": {
            "$ref": "#/components/schemas/BatchView"
          },
          "result_versions": {
            "items": {
              "$ref": "#/components/schemas/AnalysisVersionView"
            },
            "type": "array",
            "title": "Result Versions"
          },
          "freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchFreshness"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "status",
          "params",
          "results",
          "batch"
        ],
        "title": "AnalysisOutput"
      },
...
      "EventConfirmInput": {
        "properties": {
          "preview_id": {
            "type": "string",
            "minLength": 1,
            "title": "Preview Id"
          },
          "request_id": {
            "type": "string",
            "maxLength": 110,
            "minLength": 1,
            "title": "Request Id"
          },
          "security_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/EventReviewInput"
            },
            "type": "array",
            "title": "Reviews"
          },
          "revision_note": {
            "type": "string",
            "title": "Revision Note",
            "default": ""
          },
          "analyze": {
            "type": "boolean",
            "title": "Analyze",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "preview_id",
          "request_id",
          "reviews"
        ],
        "title": "EventConfirmInput"
      },
...
      "EventErrorOutput": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "EventErrorOutput"
      },
      "EventImportLimitsOutput": {
        "properties": {
          "text_characters": {
            "type": "integer",
            "title": "Text Characters",
            "default": 4000000
          },
          "text_utf8_bytes": {
            "type": "integer",
            "title": "Text Utf8 Bytes",
            "default": 16000000
          },
          "request_bytes": {
            "type": "integer",
            "title": "Request Bytes",
            "default": 50331648
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": "字符按 Unicode 码点计数；UTF-8 与 JSON 包装/转义分别占字节。预览、修订预览、确认保存使用同一请求额度。"
          }
        },
        "type": "object",
        "title": "EventImportLimitsOutput"
      },
...
      "EventPreviewInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000000,
            "minLength": 2,
            "title": "Text"
          },
          "set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Id"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "EventPreviewInput"
      },
      "EventPreviewOutput": {
        "properties": {
          "preview_id": {
            "type": "string",
            "title": "Preview Id"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomEventsImport"
              },
              {
                "$ref": "#/components/schemas/EarningsEventsImport"
              }
            ],
            "title": "Document"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Candidates"
          },
          "set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Id"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "preview_id",
          "content_hash",
          "document",
          "warnings",
          "candidates",
          "set_id",
          "expected_version",
          "events"
        ],
        "title": "EventPreviewOutput"
      },
...
      "EventPromptOutput": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "custom",
              "earnings"
            ],
            "title": "Kind"
          },
          "language": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Language"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "json_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Json Schema"
          },
          "input_schema_version": {
            "type": "string",
            "title": "Input Schema Version"
          },
          "input_example": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Example"
          },
          "input_json_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Json Schema"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "language",
          "schema_version",
          "prompt",
          "json_schema",
          "input_schema_version",
          "input_example",
          "input_json_schema"
        ],
        "title": "EventPromptOutput"
      },
...
      "EventSetsOutput": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSetSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EventSetsOutput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}